requirements.txt
beautifulsoup4==4.9.1
peewee==3.13.3
psycopg2-binary==2.8.5
requests==2.24.0
tweepy==3.8.0
dnspython==2.0.0
pymongo==3.11.2
Telethon==1.18.2
croniter==0.3.36
PyYAML==5.3.1
Pillow==8.0.1
hachoir==3.1.1
